+++ Zeal/game_functions.h
@@ -252,6 +252,8 @@ int get_show_pc_names();   // Holds value of Options -> Display -> Show PC Names
 int get_show_npc_names();  // Holds value of Options -> Display -> Show NPC Names.
 std::string class_name_short(int class_id);
 std::string class_name(int class_id);
+std::string deity_name(int deity_id);
+std::string race_name_short(int race_id);
 static constexpr int kInvalidZoneId = 0;                // get_index_from_zone_name() returns 0 if no matches.
 static constexpr int kNumZoneIds = 1000;                // 0 = invalid, 999 = last reliable entry in GameWorldData
 std::string get_full_zone_name(int zone_id);            // GameWorldData::GetFullZoneName()